start to cleanup/unify accelerate and common back-ends
