+++ mlx/backend/common/load.cpp
@@ -51,7 +51,7 @@ void load(
   }
 }
 
-void Load::eval(const std::vector<array>& inputs, array& out) {
+void Load::eval_cpu(const std::vector<array>& inputs, array& out) {
   assert(inputs.size() == 0);
   out.set_data(allocator::malloc_or_wait(out.nbytes()));
